+++ CHANGELOG.md
@@ -1,5 +1,10 @@
+## 0.1.1 `17 Sep 2025`
+
+1. Rename all "findJump" to "findJump2", to avoid confusion with commands in "Find => Jump" extension
+
 ## 0.1.0 `17 Sep 2025`
 
 1. Fork form https://github.com/usernamehw/vscode-find-jump
 2. Remove firstMatch function since I never use it
-3. Change default keymaps from "ctrl+;" to "ctrl+j" for findJump.active and add "ctrl+shift+j" for findJump.
+3. Replace original "start", "end" position configs with "activateCharStart", "activateCharEnd", "activateWordStart", "activateWordEnd", "activateCharStartSelection", "activateCharEndSelection", "activateWordStartSelection", "activateWordEndSelection" commands, to make it more flexible to use
+4. Update icon